Rename to complex_entity
